refactor: emit restored devices via a stream automatically on initialize
This refactors the device restoration logic to emit a stream of
`RestoredDevice` objects on session restoration. Each device will have a
set of metadata as well as streams for the connection state as well as
any characteristic subscriptions established in a previous session.

Each stream provides the same functionality as a standard connection or
characteristic stream and will disconnect/unsubscribe on calling
`cancel()`.